requirements.txt
dj-database-url==0.4.1
Django==1.10.7
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2
djangorestframework==3.5.3
django-filter==1.0.1
djangorestframework-gis==0.11
django-raster==0.3.1
telepot==10.5
